% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{EMANOVA}
\alias{EMANOVA}
\title{main function}
\usage{
EMANOVA(
  multi_d_mat,
  predictor,
  confonding_stat = FALSE,
  confonding_var = NA,
  r_vec
)
}
\arguments{
\item{multi_d_mat}{Input three dimensional distance matrix calculate based on microbiome OTU data.
The distance can be calculated by GUniFrac function in package GUniFrac. The
dimension of the multi_d_mat should be in n times n times number of distance,
where n represents number of samples.}

\item{predictor}{centered predictor variable}

\item{confonding_stat}{variables indicate whether confonding variables should be included}

\item{confonding_var}{confonding variables}

\item{r_vec}{vectors of r values}
}
\value{
p_value_mat matrix of p-values with different distance and different r values

final p-value
}
\description{
core function for calcualte EMANOVA p-value. Detail description of method can be found in
paper "Ensemble test for microbiome data“(currently under review).
}
\examples{
library(GUniFrac)
data("throat.otu.tab")
data("throat.tree")
data("throat.meta")

smoke_stat<-rep(0,60)
for (i in 1:60) {
  if(throat.meta$SmokingStatus[i]==throat.meta$SmokingStatus[1]){
    smoke_stat[i]<-0
  }
  else{
    smoke_stat[i]<-1
  }
}

row.sum<-apply(throat.otu.tab,1,sum)
std_out<-throat.otu.tab/row.sum

##########
unifracs = GUniFrac::GUniFrac(std_out, throat.tree, alpha = c(0,0.5,1))$unifracs
Dw  = unifracs[,,"d_1"]
#Variance adjusted weighted UniFra
Du  = unifracs[,,"d_UW"]
D0  = unifracs[,,"d_0"]
D.05= unifracs[,,"d_0.5"]
D.BC= as.matrix(vegan::vegdist(std_out, method="bray"))

dis_adaptive<-abind::abind(D.05, Du, D.BC, along = 3)
EMANOVA(dis_adaptive, smoke_stat, confonding_stat = FALSE, r_vec = c(0.125,0.25,0.5,1,2))
}
